cli: add `reset-author-on-edit` revset
It's common to create empty working-copy commits while using jj, and
currently the author timestamp for a commit is only set when it is first
created. If you create an empty commit, then don't work on a repo for a
few days, and then start working on a new feature without abandoning the
working-copy commit, the author timestamp will remain as the time the
commit was created rather than being updated to the time that work began
or finished.

This commit adds a new revset `reset-author-on-edit` which specifies
which commits should have their author information updated whenever
changes are made to them. This can be configured to fit many different
workflows. By default, empty commits with no description by the current
user have their author timestamps reset, meaning that the author
timestamp will become finalized whenever a commit is given a description
or becomes non-empty. Rebasing commits never updates the author
timestamp.