merge_tools: add merge-tool-error-status-on-conflict
Some Git merge drivers can partially resolve conflicts, leaving some
conflict markers in the output file. In that case, they exit with a
status between 1 and 127 to indicate that there was a conflict remaining
in the file. We should support this convention as well, since it will
allow many Git merge drivers to be used with Jujutsu.

https://git-scm.com/docs/gitattributes#_defining_a_custom_merge_driver