merge_tools: add merge-tool-error-status-on-conflict
Some Git merge drivers can partially resolve conflicts, leaving some
conflict markers in the output file. In this case, they exit with a
status between 1 and 127 to indicate that there was a conflict remaining
in the file. It seems useful to be able to use Git merge drivers such as
Mergiraf to resolve conflicts in Jujutsu automatically, even if they
can't resolve every conflict fully.