resolve: add --all option
If many files are conflicted, it would be nice to be able to resolve all
conflicts at once without having to run `jj resolve` multiple times.
This is especially nice for merge tools which try to automatically
resolve conflicts without user input, but it can also be used for
regular merge editors like VS Code.

The new `--all` option invokes the tool for each file, stopping
immediately if an error occurs. This means the user can cancel the merge
partway through without finishing resolving every file. In this case,
all successfully-resolved files are saved before the error is reported.