config: format divergent changes similarly to hidden commits
Since divergent changes don't necessarily always have to be resolved, it
would be nice if they weren't displayed in a scary-looking format with
red text. This commit changes the formatting of divergent changes to be
almost identical to hidden commits, except with the text "divergent"
instead of "hidden". This should allow users to make the connection that
both of these cases prevent them from using the change ID to refer to
the commit, but that both are "normal" states that a commit can be in
(i.e. there is no error that must be resolved).