rev_walk: allow filtering based on parent index ranges
I only added this for ancestors-based rev-walks, since adding it for
descendants would require more complex logic, and there's no need for it
currently.

The filtering is only used for the wanted queue but not the unwanted
queue, which means this can't be implemented using a custom index
implementation either. I think it only makes sense to filter the wanted
queue, because it's unlikely a user would want to exclude only the first
parents of a root commit for a range.