conflicts: prefer emitting snapshot for first side
Often, the diffs for the first side and second side are approximately
the same size, so we have the choice of whether we want to emit a
snapshot for the first side or for the second side. Currently, we prefer
to emit a diff for the first side when possible.

I think it would make more sense to prefer emitting snapshots instead
for the first side since the first side of the merge usually makes more
sense semantically as the snapshot. For instance, in merges, the first
side is the first parent, and in rebases, the first side is the
destination to which the diff is being applied. Therefore, I think we
should prefer to emit a snapshot for the first term unless it would
create a diff that's more than 10% larger.